Convert proxyLogins.js to TypeScript
